ProCosys task file rename
