+++ src/http/apiClients/models/fusionTasks/index.ts
@@ -1,3 +1,2 @@
 export { PimsTask } from './PimsTask';
-export { ProCoSysTask } from './ProCoSysTask';
-
+export { ProCoSysTask } from './ProCoSysTaskItem';
